migrated tester related files to tester folder
